multiple authentication without overwriting provided authentications

+++ authorization-server/src/main/java/com/ssAuthServer/authorizationserver/repository/ResourceUserRepo.java
@@ -2,11 +2,10 @@ package com.ssAuthServer.authorizationserver.repository;
 
 import com.ssAuthServer.authorizationserver.entities.ResourceUser;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
 
-public interface ResourceUserRepo extends JpaRepository<ResourceUser, Integer> {
+public interface ResourceUserRepo extends JpaRepository<ResourceUser, Long> {
 
    Optional<ResourceUser> findByEmail(String email);
